Removing "no-op" line to get rid of uninitialized warning
The removed `ns++` instruction uses the `ns` variable before
initializing it. However, the instruction can be safely and completely
removed, because it is essentially no-op here: The variable `ns` is
changed, but before it is ever read, it is initialized again to 0. The
value that it is assigned in this line is never used.